consolidate mapper methods
