+++ OrderCloud.Integrations.Payment.Stripe/Mappers/StripeRefundMapper.cs
@@ -6,7 +6,7 @@ using Stripe;
 
 namespace OrderCloud.Integrations.Payment.Stripe.Mappers
 {
-    public class StripeRefundCreateMapper
+    public class StripeRefundMapper
     {
         /// <summary>
         /// https://stripe.com/docs/api/refunds/create
@@ -16,6 +16,15 @@ namespace OrderCloud.Integrations.Payment.Stripe.Mappers
             {
                 Amount = Convert.ToInt64(transaction.Amount),
                 PaymentIntent = transaction.TransactionID
+            };
+
+        public CCTransactionResult MapRefundCreateResponse(Refund refund) => 
+            new CCTransactionResult()
+            {
+                Message = refund.Status,
+                Succeeded = refund.Status.ToLower() == "succeeded",
+                TransactionID = refund.Id,
+                Amount = refund.Amount
             };
     }
 }
