--- OrderCloud.Integrations.Payment.Stripe/Mappers/StripeResponseMapper.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace OrderCloud.Integrations.Payment.Stripe.Mappers
-{
-    public class StripeResponseMapper
-    {
-    }
-}
